Routine library maintenance

+++ src/os/windows/etcpal/os_error.c
@@ -102,7 +102,7 @@ etcpal_error_t err_winsock_to_etcpal(int wsaerror)
   }
 }
 
-etcpal_error_t err_os_to_etcpal(int error)
+etcpal_error_t err_os_to_etcpal(unsigned long error)
 {
   switch (error)
   {
